Move tools at the root of flatdata-py to prevent manipulating the sys.path in the scripts
