+++ flatdata-py/generate.py
@@ -1,0 +1,10 @@
+#!/usr/bin/env python3
+'''
+ Copyright (c) 2019 HERE Europe B.V.
+ See the LICENSE file in the root of this project for license details.
+'''
+
+import generator.app as generator
+
+if __name__ == "__main__":
+    generator.main()
